Fixed display not clearing correctly when setting goal. Increased max distance displayed

+++ display_manager.c
@@ -58,7 +58,7 @@ void displayInit(void)
 // Update the display
 void displayUpdate(deviceStateInfo_t deviceState, uint16_t secondsElapsed)
 {
-    uint16_t mTravelled = 0; // TODO: If I put this inside the case statement it won't compile. Work out why!
+    uint32_t mTravelled = 0; // TODO: If I put this inside the case statement it won't compile. Work out why!
 
     switch (deviceState.displayMode) {
     case DISPLAY_STEPS:
